chore: 워크페이지 모듈화 진행(10%)
