+++ app/lib/src/pages/work_page/functions/project_detail_page.dart
@@ -1,0 +1,68 @@
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:talk_pilot/src/models/project_model.dart';
+
+class ProjectDetailPage extends StatefulWidget {
+  final ProjectModel project;
+
+  const ProjectDetailPage({super.key, required this.project});
+
+  @override
+  State<ProjectDetailPage> createState() => _ProjectDetailPageState();
+}
+
+class _ProjectDetailPageState extends State<ProjectDetailPage> {
+  late TextEditingController _memoController;
+  late SharedPreferences _prefs;
+
+  String get _memoKey => 'memo_${widget.project.title}';
+
+  @override
+  void initState() {
+    super.initState();
+    _memoController = TextEditingController();
+    _loadSavedMemo();
+    _memoController.addListener(() {
+      _prefs.setString(_memoKey, _memoController.text);
+    });
+  }
+
+  Future<void> _loadSavedMemo() async {
+    _prefs = await SharedPreferences.getInstance();
+    final savedText = _prefs.getString(_memoKey) ?? '';
+    setState(() {
+      _memoController.text = savedText;
+    });
+  }
+
+  @override
+  void dispose() {
+    _memoController.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    final project = widget.project;
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(project.title, style: const TextStyle(color: Colors.white)),
+        backgroundColor: Colors.deepPurple,
+        iconTheme: const IconThemeData(color: Colors.white),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16),
+        child: TextField(
+          controller: _memoController,
+          maxLines: null,
+          keyboardType: TextInputType.multiline,
+          decoration: const InputDecoration(
+            hintText: '내용을 작성하세요...',
+            border: InputBorder.none,
+          ),
+          style: const TextStyle(fontSize: 16),
+        ),
+      ),
+    );
+  }
+}
